fix(build): extract shouldInjectTeamGuideMcp to avoid bun:sqlite in MCP bundle
Move shouldInjectTeamGuideMcp (which depends on ProcessConfig -> database)
to a dedicated teamGuideCapability.ts file. This prevents the esbuild MCP
server bundle from pulling in bun:sqlite through the transitive import chain
(aionMcpStdio -> teamGuidePrompt -> ProcessConfig -> database -> BunSqliteDriver).
